Resolve AGM case-collision build failure
