+++ src/middleware.ts
@@ -1,0 +1,16 @@
+import { NextResponse } from "next/server";
+import type { NextRequest } from "next/server";
+
+export function middleware(request: NextRequest) {
+  if (request.nextUrl.pathname === "/AGM") {
+    const url = request.nextUrl.clone();
+    url.pathname = "/agm";
+    return NextResponse.redirect(url, 307);
+  }
+
+  return NextResponse.next();
+}
+
+export const config = {
+  matcher: ["/AGM"],
+};
